+++ alu-web/src/decks/study/study.tsx
@@ -1,6 +1,7 @@
+import { Deck } from '../types';
 import { ReviewInstanceStudy, StudyAnswerDispatch, StudyAnswerEvent, studyAnswerReducer } from './context';
 import { StudyReviewInstances } from './review-instances';
-import { apiReviewInstanceStudy, useAsyncDispatch } from '../../lookup/lookup';
+import { apiReviewInstanceStudy, backendFetch, useAsyncDispatch, useAsyncState } from '../../lookup/lookup';
 import { useMemo } from 'react';
 
 
@@ -8,8 +9,18 @@ interface StudySkillTreeProps {
   deckId: string;
   section: string;
   isPro: 'true' | 'false';
+
+  assignmentId: never;
 }
-export default function StudySkillTree({ deckId, section, isPro }: StudySkillTreeProps) {
+interface StudyAssignmentProps {
+  assignmentId: string;
+  isPro: 'true' | 'false';
+
+  deckId: never;
+  section: never;
+}
+
+export default function Study({ deckId, section, assignmentId, isPro }: StudySkillTreeProps | StudyAssignmentProps) {
   const { studyAhead, essentialOnly } = useMemo(() => {
     const urlParams = new URLSearchParams(window.location.search);
     const studyAhead = urlParams.get('studyAhead') === 'true';
@@ -17,10 +28,17 @@ export default function StudySkillTree({ deckId, section, isPro }: StudySkillTre
 
     return { studyAhead, essentialOnly };
   }, []);
+
   const [studyData, dispatchStudyData] = useAsyncDispatch<ReviewInstanceStudy, StudyAnswerEvent>(
     apiReviewInstanceStudy,
-    [parseInt(deckId), section, studyAhead, essentialOnly],
+    [parseInt(deckId), section, parseInt(assignmentId), studyAhead, essentialOnly],
     studyAnswerReducer,
+  );
+
+  const [fetchedDeck] = useAsyncState<Deck>(
+    async () => backendFetch('GET', `teachers/assignment/${assignmentId}/get-deck/`),
+    [], undefined,
+    !deckId,
   );
 
   if (!studyData)
@@ -31,8 +49,9 @@ export default function StudySkillTree({ deckId, section, isPro }: StudySkillTre
       <StudyReviewInstances
         reviewInstances={studyData.due_for_review}
         numTotal={studyData.num_total}
-        deckId={parseInt(deckId)}
+        deckId={parseInt(deckId) ?? fetchedDeck?.id}
         section={section}
+        assignmentId={parseInt(assignmentId)}
         studyAhead={studyAhead}
         isPro={isPro.toLowerCase() === 'true'}
       />