+++ teachers/urls.py
@@ -10,7 +10,11 @@ urlpatterns = [
     )),
     path(
         'classroom/<int:classroom_id>/study/<str:sub_section>/',
-        views.classroom_assignment_study,
+        views.classroom_study_sub_section,
+    ),
+    path(
+        'classroom/<int:classroom_id>/study/assignment/<int:assignment_id>/',
+        render_basic_view('decks/study.html', context_kwargs=True)
     ),
     path(
         'classroom/<int:classroom_id>/flashcards/sections/<str:sub_section>/',